Refactoring to reduce duplication.
