+++ test/services/arrivingSensor.spec.ts
@@ -140,56 +140,5 @@ describe('ArrivingContactSensorImpl', () => {
         expect(result).toBeTruthy();
 
         contactState.verify(o => o.updateValue(CONTACT_SENSOR_OPEN), Times.Once());
-    });
-
-    it('should throw an error when not initialized on set mower metadata', () => {
-        const metadata: MowerMetadata = {
-            connected: false,
-            statusTimestamp: 1
-        };
-
-        expect(() => target.setMowerMetadata(metadata)).toThrowError();
-    });
-    
-    it('should set active status to true when connected', () => {
-        log.setup(o => o.info(It.IsAny())).returns(undefined);
-
-        const contactState = new Mock<Characteristic>();
-        const statusActive = new Mock<Characteristic>();
-        statusActive.setup(o => o.updateValue(true)).returns(statusActive.object());
-
-        const service = new Mock<Service>();
-        service.setup(o => o.getCharacteristic(Characteristic.ContactSensorState)).returns(contactState.object());
-        service.setup(o => o.getCharacteristic(Characteristic.StatusActive)).returns(statusActive.object());
-        platformAccessory.setup(o => o.getServiceById(Service.ContactSensor, 'Arriving Sensor')).returns(service.object());
-
-        target.init();
-        target.setMowerMetadata({
-            connected: true,
-            statusTimestamp: 1
-        });
-
-        statusActive.verify(o => o.updateValue(true), Times.Once());
-    });
-
-    it('should set active status to false when not connected', () => {
-        log.setup(o => o.info(It.IsAny())).returns(undefined);
-
-        const contactState = new Mock<Characteristic>();
-        const statusActive = new Mock<Characteristic>();
-        statusActive.setup(o => o.updateValue(It.IsAny())).returns(statusActive.object());
-
-        const service = new Mock<Service>();
-        service.setup(o => o.getCharacteristic(Characteristic.ContactSensorState)).returns(contactState.object());
-        service.setup(o => o.getCharacteristic(Characteristic.StatusActive)).returns(statusActive.object());
-        platformAccessory.setup(o => o.getServiceById(Service.ContactSensor, 'Arriving Sensor')).returns(service.object());
-
-        target.init();
-        target.setMowerMetadata({
-            connected: false,
-            statusTimestamp: 1
-        });
-
-        statusActive.verify(o => o.updateValue(false), Times.Once());
-    });
+    });   
 });
