+++ test/services/homebridge/abstractContactSensor.spec.ts
@@ -1,9 +1,10 @@
 import { Characteristic, Service } from 'hap-nodejs';
 import { API, HAP, PlatformAccessory } from 'homebridge';
 
-import { Mock } from 'moq.ts';
+import { It, Mock, Times } from 'moq.ts';
 import { AutomowerContext } from '../../../src/automowerAccessory';
 import { PlatformLogger } from '../../../src/diagnostics/platformLogger';
+import { MowerMetadata } from '../../../src/model';
 import { Policy } from '../../../src/services/policies/policy';
 import { ContactSensorSpy } from './contactSensorSpy';
 
@@ -33,5 +34,56 @@ describe('AbstractContactSensor', () => {
 
     it('should return undefined service when not initialized', () => {
         expect(target.getUnderlyingService()).toBeUndefined();
+    });
+
+    it('should throw an error when not initialized on set mower metadata', () => {
+        const metadata: MowerMetadata = {
+            connected: false,
+            statusTimestamp: 1
+        };
+
+        expect(() => target.setMowerMetadata(metadata)).toThrowError();
+    });
+    
+    it('should set active status to true when connected', () => {
+        log.setup(o => o.info(It.IsAny())).returns(undefined);
+
+        const contactState = new Mock<Characteristic>();
+        const statusActive = new Mock<Characteristic>();
+        statusActive.setup(o => o.updateValue(true)).returns(statusActive.object());
+
+        const service = new Mock<Service>();
+        service.setup(o => o.getCharacteristic(Characteristic.ContactSensorState)).returns(contactState.object());
+        service.setup(o => o.getCharacteristic(Characteristic.StatusActive)).returns(statusActive.object());
+        platformAccessory.setup(o => o.getServiceById(Service.ContactSensor, 'Test')).returns(service.object());
+
+        target.init();
+        target.setMowerMetadata({
+            connected: true,
+            statusTimestamp: 1
+        });
+
+        statusActive.verify(o => o.updateValue(true), Times.Once());
+    });
+
+    it('should set active status to false when not connected', () => {
+        log.setup(o => o.info(It.IsAny())).returns(undefined);
+
+        const contactState = new Mock<Characteristic>();
+        const statusActive = new Mock<Characteristic>();
+        statusActive.setup(o => o.updateValue(It.IsAny())).returns(statusActive.object());
+
+        const service = new Mock<Service>();
+        service.setup(o => o.getCharacteristic(Characteristic.ContactSensorState)).returns(contactState.object());
+        service.setup(o => o.getCharacteristic(Characteristic.StatusActive)).returns(statusActive.object());
+        platformAccessory.setup(o => o.getServiceById(Service.ContactSensor, 'Test')).returns(service.object());
+
+        target.init();
+        target.setMowerMetadata({
+            connected: false,
+            statusTimestamp: 1
+        });
+
+        statusActive.verify(o => o.updateValue(false), Times.Once());
     });
 });
